+++ bitframe-authentication/services/core/src/commonMain/kotlin/bitframe/authentication/signin/SignInServiceImpl.kt
@@ -15,6 +15,7 @@ class SignInServiceImpl(
 ) : SignInService() {
     private val scope = config.scope
     override fun signIn(credentials: SignInCredentials): Later<LoginConundrum> = scope.later {
+        validate(credentials)
         val matches = usersDao.all(where = "contacts" contains credentials.alias).await()
         if (matches.isEmpty()) throw RuntimeException("User with loginId=${credentials.alias}, not found")
         val match = matches.first()