+++ buildSrc/src/main/kotlin/versions.kt
@@ -31,7 +31,6 @@ object vers {
     object asoft {
         val foundation = "1.4.0"
         val platform = foundation
-        val access = "0.0.20"
         val contacts = "0.0.30"
         val duality = "0.0.50"
         val collections = foundation